:recycle: Refactored Tauri commands
Modularize!
