+++ src/frontend/src-tauri/src/main.rs
@@ -2,24 +2,22 @@
     all(not(debug_assertions), target_os = "windows"),
     windows_subsystem = "windows"
 )]
-#[macro_use]
-extern crate lazy_static;
-mod commands;
-mod scrape;
+mod learnatvcs;
+mod powerschool;
 
 fn main() {
     let context = tauri::generate_context!();
     tauri::Builder::default()
         .menu(tauri::Menu::os_default(&context.package_info().name))
         .invoke_handler(tauri::generate_handler![
-            commands::get_teachers,
-            commands::get_user_info,
-            commands::set_user_info,
-            commands::delete,
-            commands::data_dir_exists,
-            commands::add_teacher,
-            commands::remove_teacher,
-            scrape::scrape_plans,
+            powerschool::get_teachers,
+            powerschool::get_user_info,
+            powerschool::set_user_info,
+            powerschool::delete,
+            powerschool::data_dir_exists,
+            powerschool::add_teacher,
+            powerschool::remove_teacher,
+            learnatvcs::scrape_plans,
         ])
         .run(context)
         .expect("error while running tauri application");
